Move files to Scheduling subfolder
