+++ AElf.Kernel.Tests/Concurrency/Scheduling/UnionFindNodeTest.cs
@@ -1,7 +1,7 @@
-﻿using AElf.Kernel.Concurrency;
+﻿using AElf.Kernel.Concurrency.Scheduling;
 using Xunit;
 
-namespace AElf.Kernel.Tests.Concurrency
+namespace AElf.Kernel.Tests.Concurrency.Scheduling
 {
     public class UnionFindNodeTest
     {
