+++ AElf.Kernel/AElf.Kernel.csproj
@@ -41,5 +41,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Concurrency\Execution\Messages\" />
+    <Folder Include="Concurrency\Scheduling\" />
   </ItemGroup>
 </Project>
