+++ AElf.Kernel/Concurrency/Scheduling/Batcher.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Collections.Generic;
 
-namespace AElf.Kernel.Concurrency
+namespace AElf.Kernel.Concurrency.Scheduling
 {
 	public class Batcher : IBatcher
 	{
